refactor(provider): move driver cluster into everruns-provider
Second stage of the provider abstraction. Move the tightly-coupled LLM
driver cluster out of everruns-core into everruns-provider:

  driver_registry, error, llm_retry, stream_accumulator, stream_reconnect,
  tool_types, model, model_profiles, provider, driver_helpers,
  openai_protocol, openresponses_protocol

Plus the two shared wire types: ExecutionPhase (new execution_phase module)
and ToolResultImage (into tool_types).

The cluster is cyclic (driver_registry <-> llm_retry <-> error, and
openresponses_protocol depends on nearly all of it), so it moves as a unit.
The only edges into core's agent-loop domain were:

  - ExecutionPhase / ToolResultImage: relocated into the provider crate; core
    re-exports them at their original paths (message::ExecutionPhase,
    tools::ToolResultImage).
  - The Message/RuntimeAgent/ResolvedModel -> driver-type adapters: inverted
    into a new everruns-core `llm_conversions` module as free functions. The
    orphan rule prevents them from being `From` impls in core (both `From` and
    the driver types are foreign to core), so they are plain functions, and
    call sites (command_host, reason atom, tests) now call them directly.
  - Two doc-only links to core modules (events, atoms): de-linked.

everruns-core depends on everruns-provider and re-exports every moved module
at its original path, so the rest of the workspace is unchanged. Verified:
whole workspace compiles; provider crate lib tests pass (470).
