+++ crates/provider/src/openai_protocol.rs
