+++ crates/provider/src/stream_accumulator.rs
